+++ src/editor/CanvasPositioning.h
@@ -11,5 +11,5 @@ namespace ScenariosEditorCanvasPositioning
 	void SetCurrentScenarioCenter(float x, float y);
 	void SetSize(ImVec2 Size);
 	void SetCanvasZoomRef(float* Zoom);
-	void CenterView(int x, int y);
+	void CenterView(float x, float y);
 }
